Cleanup rake/ruby syntax

+++ Rakefile
@@ -3,36 +3,36 @@
 require 'rake'
 require 'rake/clean'
 require 'pathname'
+
 require_relative 'lib/dotfiles'
 
-# HARDCODE FOR TESTING!!
-ENV['DOTFILES_VOLUME'] = File.join(home_dir, 'dtest', 'volume')
+# Uncomment to HARDCODE FOR TESTING!!
+# ENV['DOTFILES_VOLUME'] = File.join(home_dir, 'dtest', 'volume')
 
-# Check for DOTFILES_VOLUME to override (useful for testing)
+# Check for DOTFILES_VOLUME override (useful for testing)
 VOLUME_ROOT = ENV['DOTFILES_VOLUME'] || '/'
 HOME_ROOT = ENV['DOTFILES_VOLUME'] ? File.join(VOLUME_ROOT, 'home') : home_dir
 USR_LOCAL_ROOT = File.join(VOLUME_ROOT, 'usr', 'local')
-HOMEBREW_PREFIX = File.join(VOLUME_ROOT, 'opt', 'homebrew')
 SOURCE_PATHMAP_SPEC = "%{^src/,#{HOME_ROOT}/}p"
 
 task :env do
   puts "VOLUME_ROOT=#{VOLUME_ROOT}"
   puts "HOME_ROOT=#{HOME_ROOT}"
   puts "USR_LOCAL_ROOT=#{USR_LOCAL_ROOT}"
-  puts "HOMEBREW_PREFIX=#{HOMEBREW_PREFIX}"
 end
 
 SOURCES = FileList.new do |fl|
   fl.include(`git ls-files src`.lines.map { |f| f.chomp })
 end
-
 TARGETS = SOURCES.pathmap(SOURCE_PATHMAP_SPEC)
 
-# CLEAN = FileList[]
 CLOBBER.include TARGETS
 
-task :nuke do
+task :nuke => [ :env ] do
   if ENV['DOTFILES_VOLUME']
+    raise "Root volume '/' set!" if VOLUME_ROOT == '/'
+    confirm = prompt("Confirm remove volume root #{VOLUME_ROOT}? (Y/N)", "N")
+    raise "Cancel nuke" if confirm.upcase != 'Y'
     sudo "rm -Rf #{VOLUME_ROOT}"
   end
 end
@@ -50,19 +50,19 @@ task :install => [
     :link_sources,
     :osinstall,
     'git:config',
-    'git:ignore',
-    :base_packages
+    'ssh:config',
+    'ssh:github',
+    #:base_packages
   ]
-
-task :link_sources => TARGETS
 
 directory VOLUME_ROOT
 
-directory USR_LOCAL_ROOT do
-  %w(bin lib share/man).each { |d| sh "sudo mkdir -p #{File.join(USR_LOCAL_ROOT, d)}" }
+directory USR_LOCAL_ROOT
+file USR_LOCAL_ROOT do |t|
+  %w(bin lib share/man).each { |d| sudo %(mkdir -p "#{File.join(t.name, d)}") }
 end
 
-# Symlinks for config files in src
+# Generate tasks to create symlinks for config files in src
 SOURCES.each do |src|
   file src.pathmap(SOURCE_PATHMAP_SPEC) => [ src ] do |t|
     d = t.name.pathmap('%d')
@@ -70,3 +70,5 @@ SOURCES.each do |src|
     symlink File.expand_path(t.source), t.name
   end
 end
+
+task :link_sources => TARGETS
